// src/app/services/models/user-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserDto {
  avatar?: string;
  birthday?: string;
  confirmPassword: string;
  createdAt?: string;
  email: string;
  id?: string;
  password: string;
  phone?: string;
  updatedAt?: string;
  username: string;
}